refactor(dbt): separate out `DbtEventIterator` to `dbt_event_iterator.py`
